+++ src/components/SignupPage/SignupPage.view.tsx
@@ -26,7 +26,7 @@ const SignupPageView = ({
       <ProfileForm mt="60px" errors={errors} register={register}  onSubmit={onSubmit} />
       <AdditionalInfoForm control={control} errors={errors} />
       <TermsOfServiceForm control={control} errors={errors} register={register} setValue={setValue}/>
-      <PrimaryButton mb="50px" type='submit'>회원가입 완료</PrimaryButton>
+      <PrimaryButton w="343px" h="50px" mb="50px" type='submit'>회원가입 완료</PrimaryButton>
     </Box>
   );
 };
