+++ src/components/SignupPage/_fragments/ProfileForm.tsx
@@ -1,8 +1,7 @@
 import { Box, BoxProps, Center, IconButton, Input, Text } from "@chakra-ui/react";
-import FormHelper from "@components/common/FormHelper";
+import FormHelper from "@components/common/FormHelper/FormHelper";
 import ProfileIcon from "@components/common/New/@Icons/System/Profile";
 import { FieldErrors,UseFormRegister } from "react-hook-form";
-import { FormLabelProp } from "../SignupPage.type";
 import { FormDataType } from "../_hooks/useExampleForm";
 interface ProfileFormProps extends BoxProps {
   register: UseFormRegister<FormDataType>;
@@ -23,26 +22,25 @@ const ProfileForm = ({
       <Center flexDir="column">
         <IconButton mb="40px" aria-label='Change Profile' icon={<ProfileIcon w={70} h={70} shape='Fill' />} />
       </Center>
-      <FormHelper mb="50px" label="이름" labelProps={FormLabelProp} errorText={errors.username?.message}>
+      <FormHelper mb="50px" label="이름" errorText={errors.username?.message}>
         <Input
           boxSizing='border-box' p="0px 0px 0px 20px" textStyle="text" h="40px" w="343px" border={"1px solid #1A1A1A"} borderRadius="100px"
           {...register('username')} autoComplete="off" />
       </FormHelper>
 
-      <FormHelper mb="50px" label="닉네임" labelProps={FormLabelProp} errorText={errors.nickname?.message}>
+      <FormHelper mb="50px" label="닉네임" errorText={errors.nickname?.message}>
         <Input {...register('nickname')} autoComplete="off"
           boxSizing='border-box' p="0px 0px 0px 20px" textStyle="text" h="40px" w="343px" border={"1px solid #1A1A1A"} borderRadius="100px"
         />
       </FormHelper>
 
-      <FormHelper mb="50px" label="핸드폰 번호" labelProps={FormLabelProp} errorText={errors.phone?.message}>
+      <FormHelper mb="50px" label="핸드폰 번호"  errorText={errors.phone?.message}>
         <Input flexGrow={1} {...register('phone')} autoComplete="off"
           boxSizing='border-box' p="0px 0px 0px 20px" textStyle="text" h="40px" w="343px" border={"1px solid #1A1A1A"} borderRadius="100px"
         />
       </FormHelper>
 
-      <FormHelper mb="50px" label="이메일 주소" textStyle="titleSmall"
-        labelProps={FormLabelProp}
+      <FormHelper mb="50px" label="이메일 주소"
         errorText={errors.email?.message}>
         <Input {...register('email')} autoComplete="off"
           boxSizing='border-box' p="0px 0px 0px 20px" textStyle="text" h="40px" w="343px" border={"1px solid #1A1A1A"} borderRadius="100px"
