+++ src/components/common/FormHelper/FormHelper.props.tsx
@@ -1,0 +1,16 @@
+import { FormLabelProps, FormErrorMessageProps } from "@chakra-ui/react";
+
+export const formLabelProp: FormLabelProps = {
+  marginBottom: "10px",
+  color: "primary.500",
+  fontSize: "12px",
+  lineHeight: "18px",
+  fontWeight: 700,
+}
+export const errorTextProp:FormErrorMessageProps = {
+  fontSize: "12px",
+  lineHeight: "18px",
+  fontWeight: 400,
+  mt:"10px",
+  color:"alertWarning",
+};
